chore: enable unit_test_file_layout and unicode_ellipsis_in_docs scan_code_spans
Turn on both rules for the crate's own self-lint and make the source
satisfy them (building on the separate-target fix already on master):

- Drop the crate's `unit_test_file_layout` opt-out and extract the eight
  over-budget inline `#[cfg(test)] mod tests` blocks under `src/` into
  nested `tests.rs` files.
- Set `unicode_ellipsis_in_docs.scan_code_spans = true` and reword the
  rules' own doc comments so no `…` glyph survives inside an inline code
  span; regenerate the rule catalogue.
- Pin the `ui/` default-config sweep to an empty `dylint.toml` so the
  crate-wide `scan_code_spans` override doesn't leak into it (the
  enabled behaviour stays covered under `ui-toml/`).
