+++ rules/unicode_ellipsis_in_docs.md
@@ -8,7 +8,7 @@
 > U+2026 HORIZONTAL ELLIPSIS in doc comments; prefer `...`
 
 ## What it does
-Forbids U+2026 HORIZONTAL ELLIPSIS (`…`) in doc comments —
+Forbids U+2026 HORIZONTAL ELLIPSIS in doc comments —
 `///` and `//!` line forms and the `/** */` / `/*! */` block
 forms. Prefer the three-ASCII-dot form `...`. Regular `//` and
 `/* */` comments are covered by a sibling lint
@@ -19,10 +19,10 @@ This is a stylistic preference, not a correctness issue.
 ASCII `...` survives every encoding round-trip, every terminal,
 every copy-paste, every `grep` invocation, and every `git diff`
 viewer without rendering as `?` or a tofu box. The visual
-difference between `…` and `...` is small enough that the
-Unicode form usually arrives by accident — autocorrect, an IDE
-smart-quote setting — rather than as a deliberate choice in
-technical writing.
+difference between the Unicode form and `...` is small enough
+that the Unicode form usually arrives by accident — autocorrect,
+an IDE smart-quote setting — rather than as a deliberate choice
+in technical writing.
 
 ## Example
 ```rust,ignore
@@ -46,10 +46,9 @@ pipelines occasionally insert. Empty by default.
 
 ### `scan_code_spans`: `boolean` (optional)
 
-Whether to also flag a character inside an inline code span
-(`` `…` ``). Defaults to `false`: code spans often quote example
-text where the ellipsis is meaningful, so they are left alone
-unless this is set to `true`. Code *blocks* — fenced
-(` ``` … ``` `), `~~~`-fenced, four-space indented, and the
-doc-test code they hold — are always skipped regardless of this
-knob.
+Whether to also flag a character inside an inline code span.
+Defaults to `false`: code spans often quote example text where
+the ellipsis is meaningful, so they are left alone unless this
+is set to `true`. Code *blocks* — fenced (` ``` ... ``` `),
+`~~~`-fenced, four-space indented, and the doc-test code they
+hold — are always skipped regardless of this knob.
