+++ src/rules/unicode_ellipsis_in_docs.rs
@@ -9,7 +9,7 @@ use crate::markdown::{position_in_skip, scan_code_regions};
 
 declare_tool_lint! {
     /// ### What it does
-    /// Forbids U+2026 HORIZONTAL ELLIPSIS (`…`) in doc comments —
+    /// Forbids U+2026 HORIZONTAL ELLIPSIS in doc comments —
     /// `///` and `//!` line forms and the `/** */` / `/*! */` block
     /// forms. Prefer the three-ASCII-dot form `...`. Regular `//` and
     /// `/* */` comments are covered by a sibling lint
@@ -20,10 +20,10 @@ declare_tool_lint! {
     /// ASCII `...` survives every encoding round-trip, every terminal,
     /// every copy-paste, every `grep` invocation, and every `git diff`
     /// viewer without rendering as `?` or a tofu box. The visual
-    /// difference between `…` and `...` is small enough that the
-    /// Unicode form usually arrives by accident — autocorrect, an IDE
-    /// smart-quote setting — rather than as a deliberate choice in
-    /// technical writing.
+    /// difference between the Unicode form and `...` is small enough
+    /// that the Unicode form usually arrives by accident — autocorrect,
+    /// an IDE smart-quote setting — rather than as a deliberate choice
+    /// in technical writing.
     ///
     /// ### Example
     /// ```rust,ignore
@@ -49,13 +49,12 @@ struct Config {
     /// or U+2025 TWO DOT LEADER (`‥`) that the same autocorrect
     /// pipelines occasionally insert. Empty by default.
     extra_flagged_chars: Vec<char>,
-    /// Whether to also flag a character inside an inline code span
-    /// (`` `…` ``). Defaults to `false`: code spans often quote example
-    /// text where the ellipsis is meaningful, so they are left alone
-    /// unless this is set to `true`. Code *blocks* — fenced
-    /// (` ``` … ``` `), `~~~`-fenced, four-space indented, and the
-    /// doc-test code they hold — are always skipped regardless of this
-    /// knob.
+    /// Whether to also flag a character inside an inline code span.
+    /// Defaults to `false`: code spans often quote example text where
+    /// the ellipsis is meaningful, so they are left alone unless this
+    /// is set to `true`. Code *blocks* — fenced (` ``` ... ``` `),
+    /// `~~~`-fenced, four-space indented, and the doc-test code they
+    /// hold — are always skipped regardless of this knob.
     scan_code_spans: bool,
 }
